SourceCode/requirements.txt
Pillow==11.0.0
auto-py-to-exe==2.42.0 